requirements.txt
django-mptt==0.16.0
factory-boy==2.11.1
mock==2.0.0
pyhamcrest==1.9.0
django-cte==1.1.4
django-codemirror2==0.2
behave-django==1.3.0